requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
ipython>=8.12.0
ollama>=0.1.4 